refactor(annotation-server): structure common misc into common utils
