+++ annotation-server/src/guidelines/caches/guideline-cache.ts
@@ -1,4 +1,4 @@
-import { CacheMap } from '../../common/cache-map';
+import { CacheMap } from '../../common/utils/cache-map';
 import { GuidelineError } from '../entities/guideline-error.entity';
 
 export abstract class GuidelineCacheMap<Value> extends CacheMap<
